+++ src/Colors.js
@@ -1,0 +1,15 @@
+const base = {
+  textLight: 'white'
+};
+
+const Colors = {
+  navButtonBackground: 'hsla(0, 0%, 50%, 0.2)',
+  navButtonBackgroundHover: 'hsla(0, 0%, 50%, 0.50)',
+  navButtonTextColor: 'hsla(0, 0%, 10%, 0.30)',
+  navButtonTextColorHover: 'hsla(0, 0%, 100%, 1.0)',
+  infoPanelBackground: 'hsla(0, 0%, 10%, 0.3)',
+  infoPanelText: base.textLight,
+  infoPanelLink: 'hsla(0, 0%, 30%, 1)'
+};
+
+export default Colors;